requirements.txt
streamlit
pandas
pyarrow
azure-storage-file-datalake
sqlalchemy
pyodbc
python-dotenv
adlfs
